fix(tests): fix developer password in tests since use of secret token
affects: @zetapush/cli, @zetapush/client, @zetapush/common, @zetapush/integration, @zetapush/testing
